requirements.txt
opencolorio==2.1.0
colour-science==0.3.16